requirements.txt
# Google Agent Development Kit
google-adk>=0.1.0

# Google Generative AI
google-generativeai>=0.8.0

# Google Cloud Vertex AI
google-cloud-aiplatform>=1.38.0
